Included manufacturing test
As we talked about in class today, complications can exist in these types of optimization problems that don't play nicely with derivative-based methods. In this example, I added in a variation on the manufactured part (triangular distribution) and required that 99% of the simluated parts meet this standard. This changes the type of answer we get. The produced answer meets the criteria in the face of the new requirements, and would be much harder to get by hand.